Deprecating all C headers
