+++ doc/Tutorials/Approximate-Synchronizer-Cpp.rst
@@ -18,9 +18,9 @@ If you have not done so already `create a workspace <https://docs.ros.org/en/rol
   #include <functional>
   #include <memory>
 
-  #include "message_filters/subscriber.h"
-  #include "message_filters/synchronizer.h"
-  #include "message_filters/sync_policies/approximate_time.h"
+  #include "message_filters/subscriber.hpp"
+  #include "message_filters/synchronizer.hpp"
+  #include "message_filters/sync_policies/approximate_time.hpp"
 
   #include "sensor_msgs/msg/temperature.hpp"
   #include "sensor_msgs/msg/fluid_pressure.hpp"
