+++ include/message_filters/null_types.h
@@ -1,76 +1,22 @@
-// Copyright 2010, Willow Garage, Inc. All rights reserved.
+// Copyright 2024 Open Source Robotics Foundation, Inc.
 //
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-//    * Redistributions of source code must retain the above copyright
-//      notice, this list of conditions and the following disclaimer.
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
-//    * Redistributions in binary form must reproduce the above copyright
-//      notice, this list of conditions and the following disclaimer in the
-//      documentation and/or other materials provided with the distribution.
-//
-//    * Neither the name of the Willow Garage nor the names of its
-//      contributors may be used to endorse or promote products derived from
-//      this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #ifndef MESSAGE_FILTERS__NULL_TYPES_H_
 #define MESSAGE_FILTERS__NULL_TYPES_H_
 
-#include <memory>
+#warning This header is obsolete, please include message_filters/null_types.hpp instead
 
-#include <rclcpp/rclcpp.hpp>
-
-#include "message_filters/connection.h"
-#include "message_filters/message_traits.h"
-
-namespace message_filters
-{
-
-struct NullType
-{
-};
-typedef std::shared_ptr<NullType const> NullTypeConstPtr;
-
-template<class M>
-struct NullFilter
-{
-  template<typename C>
-  Connection registerCallback(const C &)
-  {
-    return Connection();
-  }
-
-  template<typename P>
-  Connection registerCallback(const std::function<void(P)> &)
-  {
-    return Connection();
-  }
-};
-
-namespace message_traits
-{
-template<>
-struct TimeStamp<message_filters::NullType>
-{
-  static rclcpp::Time value(const message_filters::NullType &)
-  {
-    return rclcpp::Time();
-  }
-};
-}  // namespace message_traits
-}  // namespace message_filters
+#include <message_filters/null_types.hpp>
 
 #endif  // MESSAGE_FILTERS__NULL_TYPES_H_
