+++ include/message_filters/pass_through.h
@@ -1,94 +1,22 @@
-// Copyright 2010, Willow Garage, Inc. All rights reserved.
+// Copyright 2024 Open Source Robotics Foundation, Inc.
 //
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-//    * Redistributions of source code must retain the above copyright
-//      notice, this list of conditions and the following disclaimer.
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
-//    * Redistributions in binary form must reproduce the above copyright
-//      notice, this list of conditions and the following disclaimer in the
-//      documentation and/or other materials provided with the distribution.
-//
-//    * Neither the name of the Willow Garage nor the names of its
-//      contributors may be used to endorse or promote products derived from
-//      this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #ifndef MESSAGE_FILTERS__PASS_THROUGH_H_
 #define MESSAGE_FILTERS__PASS_THROUGH_H_
 
-#include <memory>
-#include <utility>
-#include <vector>
+#warning This header is obsolete, please include message_filters/pass_through.hpp instead
 
-#include "message_filters/simple_filter.h"
-
-namespace message_filters
-{
-/**
- * \brief Simple passthrough filter.  What comes in goes out immediately.
- */
-template<typename M>
-class PassThrough : public SimpleFilter<M>
-{
-public:
-  typedef std::shared_ptr<M const> MConstPtr;
-  typedef MessageEvent<M const> EventType;
-
-  PassThrough()
-  {
-  }
-
-
-  template<typename F>
-  PassThrough(F & f)  // NOLINT(runtime/explicit)
-  {
-    connectInput(f);
-  }
-
-  template<class F>
-  void connectInput(F & f)
-  {
-    incoming_connection_.disconnect();
-    incoming_connection_ =
-      f.registerCallback(
-      typename SimpleFilter<M>::EventCallback(
-        std::bind(
-          &PassThrough::cb, this,
-          std::placeholders::_1)));
-  }
-
-  void add(const MConstPtr & msg)
-  {
-    add(EventType(msg));
-  }
-
-  void add(const EventType & evt)
-  {
-    this->signalMessage(evt);
-  }
-
-private:
-  void cb(const EventType & evt)
-  {
-    add(evt);
-  }
-
-  Connection incoming_connection_;
-};
-
-}  // namespace message_filters
+#include <message_filters/pass_through.hpp>
 
 #endif  // MESSAGE_FILTERS__PASS_THROUGH_H_
