+++ src/connection.cpp
@@ -26,7 +26,7 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#include "message_filters/connection.h"
+#include "message_filters/connection.hpp"
 
 namespace message_filters
 {
