+++ test/test_approximate_epsilon_time_policy.cpp
@@ -36,9 +36,9 @@
 
 #include <rclcpp/rclcpp.hpp>
 
-#include "message_filters/synchronizer.h"
-#include "message_filters/sync_policies/approximate_epsilon_time.h"
-#include "message_filters/message_traits.h"
+#include "message_filters/synchronizer.hpp"
+#include "message_filters/sync_policies/approximate_epsilon_time.hpp"
+#include "message_filters/message_traits.hpp"
 
 struct Header
 {
