+++ test/test_approximate_time_policy.cpp
@@ -34,9 +34,9 @@
 #include <vector>
 
 #include <rclcpp/rclcpp.hpp>
-#include "message_filters/synchronizer.h"
-#include "message_filters/sync_policies/approximate_time.h"
-#include "message_filters/message_traits.h"
+#include "message_filters/synchronizer.hpp"
+#include "message_filters/sync_policies/approximate_time.hpp"
+#include "message_filters/message_traits.hpp"
 
 struct Header
 {
