+++ test/test_exact_time_policy.cpp
@@ -32,8 +32,8 @@
 #include <memory>
 
 #include <rclcpp/rclcpp.hpp>
-#include "message_filters/synchronizer.h"
-#include "message_filters/sync_policies/exact_time.h"
+#include "message_filters/synchronizer.hpp"
+#include "message_filters/sync_policies/exact_time.hpp"
 
 struct Header
 {
