+++ test/test_fuzz.cpp
@@ -36,10 +36,10 @@
 #include <random>
 
 #include <rclcpp/rclcpp.hpp>
-#include "message_filters/subscriber.h"
-#include "message_filters/time_sequencer.h"
-#include "message_filters/time_synchronizer.h"
-#include "message_filters/chain.h"
+#include "message_filters/subscriber.hpp"
+#include "message_filters/time_sequencer.hpp"
+#include "message_filters/time_synchronizer.hpp"
+#include "message_filters/chain.hpp"
 #include "sensor_msgs/msg/imu.hpp"
 
 typedef sensor_msgs::msg::Imu Msg;
